requirements.txt
flask
gunicorn
flask-cors
python-dotenv
waitress
